+++ packages/hooks/src/useMessage.tsx
@@ -159,4 +159,4 @@ export function useMessage() {
 }
 
 
-export default useMessage;
+// export default useMessage;