+++ packages/utils/src/networks/axios/axiosRetry.ts
@@ -1,4 +1,5 @@
 import { AxiosError, AxiosInstance } from 'axios';
+import type { RetryInternalAxiosRequestConfig } from "@viaz/types";
 /**
  *  请求重试机制
  */
@@ -9,14 +10,17 @@ export class AxiosRetry {
    */
   retry(axiosInstance: AxiosInstance, error: AxiosError) {
     // @ts-ignore
-    const { config } = error.response;
+    // const { config  }         = error.response;
+    const config = error.response?.config as RetryInternalAxiosRequestConfig;
+    // @ts-ignore
     const { waitTime, count } = config?.requestOptions?.retryRequest;
     config.__retryCount = config.__retryCount || 0;
     if (config.__retryCount >= count) {
       return Promise.reject(error);
     }
     config.__retryCount += 1;
-    //请求返回后config的header不正确造成重试请求失败,删除返回headers采用默认headers
+      //请求返回后config的header不正确造成重试请求失败,删除返回headers采用默认headers
+    // @ts-ignore
     delete config.headers;
     return this.delay(waitTime).then(() => axiosInstance(config));
   }