+++ packages/utils/src/src/index.ts
@@ -1,6 +1,0 @@
-export * from "./vue";
-export * from "./networks";
-export * from "./is";
-export * from "./env";
-export * from "./log";
-export * from "./oths";