+++ packages/directives/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @viaz/directives
 
+## 0.0.2-beta.12
+
+### Patch Changes
+
+- fixed build ts error
+- Updated dependencies
+  - @viaz/types@0.0.2-beta.12
+
 ## 0.0.2-beta.11
 
 ### Patch Changes