+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @viaz/utils
 
+## 0.0.2-beta.12
+
+### Patch Changes
+
+- fixed build ts error
+- Updated dependencies
+  - @viaz/constants@0.0.2-beta.12
+  - @viaz/enums@0.0.2-beta.12
+  - @viaz/hooks@0.0.2-beta.12
+  - @viaz/types@0.0.2-beta.12
+
 ## 0.0.2-beta.11
 
 ### Patch Changes