+++ packages/viaz/CHANGELOG.md
@@ -1,5 +1,18 @@
 # viaz
 
+## 0.0.2-beta.12
+
+### Patch Changes
+
+- fixed build ts error
+- Updated dependencies
+  - @viaz/components@0.0.2-beta.12
+  - @viaz/directives@0.0.2-beta.12
+  - @viaz/constants@0.0.2-beta.12
+  - @viaz/hooks@0.0.2-beta.12
+  - @viaz/types@0.0.2-beta.12
+  - @viaz/utils@0.0.2-beta.12
+
 ## 0.0.2-beta.11
 
 ### Patch Changes